+++ requirements.txt
@@ -4,7 +4,7 @@ et-xmlfile==1.0.1
 jdcal==1.2
 openpyxl==2.3.2
 psycopg2==2.6.1
-reportlab==3.2.0
-selenium==2.48.0
-django-jenkins==0.18.1
-coverage==4.0.3
+#reportlab==3.2.0
+#selenium==2.48.0
+#django-jenkins==0.18.1
+#coverage==4.0.3
